+++ ukis_pysat/download.py
@@ -8,15 +8,18 @@ import traceback
 
 import fiona
 import landsatxplore.api
+import numpy as np
 import pyproj
 import requests
 import sentinelsat
-from matplotlib.pyplot import imread
+from io import BytesIO
+from PIL import Image
 from pylandsat import Product
 from shapely import geometry, wkt, ops
 
 from ukis_pysat.members import Datahub, Platform
 from ukis_pysat.file import env_get, pack
+
 
 logger = logging.getLogger(__name__)
 
@@ -100,9 +103,9 @@ class Source:
         :param platform: image platform (<enum 'Platform'>).
         :param date: Date from - to (String or Datetime tuple). Expects a tuple of (start, end), e.g.
             (yyyyMMdd, yyyy-MM-ddThh:mm:ssZ, NOW, NOW-<n>DAY(S), HOUR(S), MONTH(S), etc.)
-        :param aoi: Area of interest as GeoJson file or bounding box tuple (String, Tuple).
+        :param aoi: Area of interest as GeoJson file or bounding box tuple with lat lon coordinates (String, Tuple).
         :param cloud_cover: Percent cloud cover scene from - to (Integer tuple).
-        :returns: Metadata of products that match query criteria (GeoJSON-like mapping).
+        :returns: Metadata of products that match query criteria (List of GeoJSON-like mappings).
         """
         if self.src == Datahub.file:
             raise NotImplementedError("File metadata query not yet supported.")
@@ -135,7 +138,6 @@ class Source:
                 if cloud_cover and platform != platform.Sentinel1:
                     kwargs["cloudcoverpercentage"] = cloud_cover
                 meta_src = self.api.query(
-                    # area=sentinelsat.geojson_to_wkt(sentinelsat.read_geojson(aoi)),
                     area=self._prep_aoi(aoi).wkt,
                     date=date,
                     platformname=platform.value,
@@ -149,9 +151,11 @@ class Source:
                 )
 
         try:
-            # construct harmonized metadata
+            # construct list of harmonized metadata
+            meta = []
             for m in meta_src:
-                return self.construct_metadata(meta_src=m)
+                meta.append(self.construct_metadata(meta_src=m))
+            return meta
         except Exception as e:
             raise Exception(
                 f"{traceback.format_exc()} Could not convert metadata to custom metadata. "
@@ -228,31 +232,35 @@ class Source:
 
         return geometry.mapping(_GeoInterface({"type": "Feature", "properties": prop, "geometry": geom}))
 
-    def get_metadata(self, product_id):
-        """This method gets satellite image metadata for a specific product from a local file directory.
+    def filter_metadata(self, meta, filter_dict):
+        """This method filters metadata as returned by query_metadata() based on filter_dict.
 
-        :param product_id: ID of the satellite image product (String).
-        :returns: Metadata of satellite product that matches the product_id (GeoJson).
+        :param meta: Metadata of product(s) (List of GeoJSON-like mappings).
+        :param filter_dict: Key value pair to use as filter e.g. {"producttype": "S2MSI1C"} (Dictionary).
+        :returns: Metadata of products that match filter criteria (List of GeoJSON-like mappings).
         """
-        if self.src == Datahub.file:
-            # search local directory for metadata of a particular image
-            for root, dirs, files in os.walk(str(self.api)):
-                for file in files:
-                    if file.endswith(".json") and product_id in file:
-                        with open(os.path.join(root, file)) as json_file:
-                            return json.load(json_file)
+        m = []
+        k = list(filter_dict.keys())
+        for i in range(len(meta)):
+            if filter_dict[k[0]] == meta[i]["properties"][k[0]]:
+                m.append(meta[i])
+        return m
 
-        elif self.src == Datahub.EarthExplorer:
-            raise NotImplementedError(f"get_metadata not supported for {self.src}.")
+    def download_metadata(self, meta, target_dir):
+        """This method writes metadata as returned by query_metadata() to file.
 
-        elif self.src == Datahub.Scihub:
-            raise NotImplementedError(f"get_metadata not supported for {self.src}.")
+        :param metadata: Metadata of product(s) (GeoJSON-like mapping).
+        :param target_dir: Target directory that holds the downloaded metadata (String)
+        """
+        for i in range(len(meta)):
+            with open(os.path.join(target_dir, meta[i]["properties"]["srcid"] + ".json"), "w") as f:
+                json.dump(meta[i], f)
 
-    def download_image(self, product_srcid, product_uuid, target_dir):
+    def download_image(self, platform, product_uuid, target_dir):
         """This method downloads satellite image data to a target directory for a specific product_id.
         Incomplete downloads are continued and complete files are skipped.
 
-        :param product_srcid: Product source id (String).
+        :param platform: image platform (<enum 'Platform'>).
         :param product_uuid: UUID of the satellite image product (String).
         :param target_dir: Target directory that holds the downloaded images (String)
         """
@@ -261,9 +269,12 @@ class Source:
 
         elif self.src == Datahub.EarthExplorer:
             try:
-                # NOTE: this downloads data from Amazon AWS using pylandsat. landsatxplore is great for metadata
-                # search on EE but download via EE is slow. pylandsat is great for fast download from AWS but is
-                # poor for metadata search. Here we combine them to get the best from both packages.
+                # query EarthExplorer for srcid of product
+                meta_src = self.api.request("metadata", **{"datasetName": platform.value, "entityIds": [product_uuid],},)
+                product_srcid = meta_src[0]["displayId"]
+                # download data from AWS
+                # landsatxplore is great for metadata search on EE but download via EE is slow. pylandsat is great
+                # for fast download from AWS but does not provide good metadata search. Here we combine their benefits.
                 product = Product(product_srcid)
                 product.download(out_dir=target_dir, progressbar=False)
             except Exception as e:
@@ -292,24 +303,24 @@ class Source:
                     f"This Exception was raised: {e}."
                 )
 
-    def download_quicklook(self, platform, product_uuid, product_srcid, target_dir):
+    def download_quicklook(self, platform, product_uuid, target_dir):
         """This method downloads a quicklook of the satellite image to a target directory for a specific product_id.
+        It performs a very rough geocoding of the quicklooks by shifting the image to the location of the footprint.
 
         :param platform: image platform (<enum 'Platform'>).
         :param product_uuid: UUID of the satellite image product (String).
-        :param product_srcid: Product source id (String).
         :param target_dir: Target directory that holds the downloaded images (String)
         """
         if self.src == Datahub.file:
-            logger.warning(f"download_quicklook not supported for {self.src}.")
-            return
+            raise NotImplementedError(f"download_quicklook not supported for {self.src}.")
 
         elif self.src == Datahub.EarthExplorer:
             try:
-                m = self.api.request("metadata", **{"datasetName": platform.value, "entityIds": [product_uuid],},)
-                url = m[0]["browseUrl"]
-                bounds = geometry.shape(m[0]["spatialFootprint"]).bounds
-                self.save_quicklook_image(url, bounds, product_srcid, target_dir)
+                # query EarthExplorer for url, srcid and bounds of product
+                meta_src = self.api.request("metadata", **{"datasetName": platform.value, "entityIds": [product_uuid],},)
+                url = meta_src[0]["browseUrl"]
+                bounds = geometry.shape(meta_src[0]["spatialFootprint"]).bounds
+                product_srcid = meta_src[0]["displayId"]
             except Exception as e:
                 logger.warning(
                     f"{traceback.format_exc()} Could not download and save quicklook. "
@@ -318,32 +329,28 @@ class Source:
 
         elif self.src == Datahub.Scihub:
             try:
-                m = self.api.get_product_odata(product_uuid)
+                # query Scihub for url, srcid and bounds of product
+                meta_src = self.api.get_product_odata(product_uuid)
                 url = "https://scihub.copernicus.eu/apihub/odata/v1/Products('{}')/Products('Quicklook')/$value".format(
                     product_uuid
                 )
-                bounds = wkt.loads(m["footprint"]).bounds
-                self.save_quicklook_image(url, bounds, product_srcid, target_dir)
+                bounds = wkt.loads(meta_src["footprint"]).bounds
+                product_srcid = meta_src["title"]
             except Exception as e:
                 logger.warning(
                     f"{traceback.format_exc()} Could not download and save quicklook. This Exception was raised: {e}."
                 )
 
-    def save_quicklook_image(self, platform, bounds, product_srcid, target_dir):
-        """This method saves a quicklook of the satellite image to a target directory for a specific product_id.
+        # download quicklook and crop no-data borders
+        response = requests.get(url, auth=(self.user, self.pw))
+        quicklook = np.asarray(Image.open(BytesIO(response.content)))
+        # use threshold of 50 to overcome noise in JPEG compression
+        xs, ys, zs = np.where(quicklook >= 50)
+        quicklook = quicklook[min(xs):max(xs)+1, min(ys):max(ys)+1, min(zs):max(zs)+1]
+        Image.fromarray(quicklook).save(os.path.join(target_dir, product_srcid + ".jpg"))
 
-        :param platform: image platform (<enum 'Platform'>).
-        :param bounds: Bounding Box of footprint
-        :param product_srcid: Product source id (String).
-        :param target_dir: Target directory that holds the downloaded images (String)
-        """
-        response = requests.get(platform, auth=(self.user, self.pw))
-        with open(os.path.join(target_dir, product_srcid + ".jpg"), "wb") as f:
-            f.write(response.content)
-        quicklook = imread(os.path.join(target_dir, product_srcid + ".jpg"))
+        # geocode quicklook
         quicklook_size = (quicklook.shape[1], quicklook.shape[0])
-
-        # save worldfile
         dist_x = geometry.Point(bounds[0], bounds[1]).distance(geometry.Point(bounds[2], bounds[1])) / quicklook_size[0]
         dist_y = geometry.Point(bounds[0], bounds[1]).distance(geometry.Point(bounds[0], bounds[3])) / quicklook_size[1]
         ul_x, ul_y = bounds[0], bounds[3]