+++ CHANGELOG.rst
@@ -21,16 +21,22 @@ Added
 - ``download.Source()``: Static method ``_prep_aoi()`` for on the fly perparation of aoi for queries #1
 - ``download.Source()``: ``query()`` accepts now aoi in forms of geojson file with varying CRS or bounding box coordinates in Lat Lon #1
 - requirements: pyproj #1
+- ``download.Source()``: added methods to filter and download metadata #4
+- Sentinel3 test #10
 
 Fixed
 *****
+- download.Source(): Improved geocoding quicklooks #5
 - fixed #7
 
 Changed
 *******
+
 - ``download.Source()``: Moved all metadata mapping from ``query()`` to ``construct_metadata()`` #1
 - ``download.Source()``: Changed ``_construct_metadata()`` to ``construct_metadata()`` and removed static #1
 - ``download.Source()``: Simplified api queries in ``query()`` #1
+- ``download.Source()``: removed ``get_metadata()`` #4
+- requirements: Removed matplotlib #9
 
 [0.1.0] – 2020-04-29
 --------------------